integrate with current input modal
