+++ src/extension/features/accounts/auto-reconcile/styles.scss
@@ -1,23 +1,7 @@
-.tk-auto-reconcile-modal {
-  width: 300px;
-}
-
-.tk-auto-reconcile-input {
-  width: 100%;
-  border-radius: 0.4em;
-  border: 2px solid var(--textinput_border_active);
-  margin-bottom: 0.5rem;
-}
-
 .tk-autoreconcile-text {
   font-size: 0.9em;
   font-weight: normal;
   text-align: center;
-}
-
-.tk-auto-reconcile-error {
-  color: red;
-  margin-bottom: 0.5rem;
 }
 
 .tk-modal-stack {
